requirements.txt
flask==2.3.2
requests==2.31.0
python-networkmanager==2.2
pyyaml==6.0
pynacl==1.5.0
itsdangerous==2.1.2
apscheduler==3.10.4
pytz==2023.3
pytest==7.4.4
pytest-mock==3.12.0
freezegun==1.3.0
